basicwidget: add ranged font weights via the wght variation axis
Add the first metric-affecting ranged styles: per-range OpenType
variations and features, with font weight modeled as the wght variation
axis rather than a dedicated style property. The core exposes
SetVariationInRange/UnsetVariationInRange and SetFeatureInRange/
UnsetFeatureInRange; basicwidget.Text exposes SetWeightInRange as a
wght wrapper.

textutil gains FaceRun and Style.FaceRuns: per-range face overrides,
sorted and disjoint, measured and drawn as standalone segments split at
tab and face boundaries (advanceWithFaces). Wrap, measure, hit-testing,
caret geometry, and drawing honor the face runs; styled drawing is
unified in drawStyledVisualLine, whose tab- and face-delimited segments
are placed to match advanceWithFaces while color chunks within a
segment keep the shaping context. The single-face fast paths — the
content-keyed layout cache, the per-logical-line measurer, the viewport
walker, and glyph-based hit-testing — refuse when face runs are present
and fall back to run-aware whole-document paths, so text without metric
overrides renders byte-identically as before.

textstyle owns the ranged styles' serialization: Runs.WriteStateKey and
Runs.WriteMetricStateKey write through a small Writer interface that
guigui.StateKeyWriter satisfies, and the widget fingerprints the metric
properties with an FNV adapter to reset size caches when a metric
override changes. The widget resolves faces per run from the overrides'
variations and features, appends them into a caller-cleared buffer, and
skips face runs while masking or during an IME composition. The
gallery's Rich Texts section replaces the bold toggle with inline bold
ranges.

Updates #131

+++ basicwidget/internal/textwidget/statekey.go
@@ -18,10 +18,10 @@ func writeColor(w *guigui.StateKeyWriter, c color.Color) {
 	}
 	w.WriteBool(true)
 	r, g, b, a := c.RGBA()
-	w.WriteUint32(r)
-	w.WriteUint32(g)
-	w.WriteUint32(b)
-	w.WriteUint32(a)
+	w.WriteUint16(uint16(r))
+	w.WriteUint16(uint16(g))
+	w.WriteUint16(uint16(b))
+	w.WriteUint16(uint16(a))
 }
 
 // writePadding writes a guigui.Padding into w.
